feature: chain adapter
