+++ apps/frontend/src/components/admin/index.ts
@@ -1,0 +1,6 @@
+export { default as AdminDashboard } from "./AdminDashboard";
+export { ActivityCard } from "./ActivityCard";
+export { QRCodeSection } from "./QRCodeSection";
+export { default as QRCodeDisplay } from "./QRCodeDisplay";
+export { default as RevealKeyDialog } from "./RevealKeyDialog";
+export { StatsCards } from "./StatsCards";
