+++ apps/frontend/src/components/claim/index.ts
@@ -1,0 +1,5 @@
+export { ActionPanel } from "./ActionPanel";
+export { default as ClaimForm } from "./ClaimForm";
+export { default as ClaimTicketPanel } from "./ClaimTicketPanel";
+export { ManualInputPanel } from "./ManualInputPanel";
+export { default as QRCodeScanner } from "./QRCodeScanner";
