+++ apps/frontend/src/components/layout/ClientProviders.tsx
@@ -2,13 +2,12 @@
 
 import React from "react";
 import dynamic from "next/dynamic";
-import { Header } from "@/components/layout/Header";
-import { Toaster } from "@/components/ui/toaster";
-import { TrpcProvider } from "@/components/layout/TrpcProvider";
+import { Header, TrpcProvider } from "@/components/layout";
+import { Toaster } from "@/components/ui";
 
 // Load RainbowKit and Wagmi providers only on the client to avoid SSR IndexedDB access
 const RainbowKitProvider = dynamic(
-  () => import("@/components/layout/RainbowKitProvider").then((m) => m.RainbowKitProvider),
+  () => import("@/components/layout").then((m) => m.RainbowKitProvider),
   { ssr: false }
 );
