+++ apps/frontend/src/components/layout/brand/index.ts
@@ -1,0 +1,1 @@
+export { ZkAssetRaffleLogo } from "./ZkAssetRaffleLogo";
