+++ apps/frontend/src/components/layout/index.ts
@@ -1,0 +1,5 @@
+export { ClientProviders } from "./ClientProviders";
+export { Header } from "./Header";
+export { RainbowKitProvider } from "./RainbowKitProvider";
+export { TrpcProvider } from "./TrpcProvider";
+export { ZkAssetRaffleLogo } from "./brand/ZkAssetRaffleLogo";
